refactor: rename TaskPriority.tsx to TaskPriorityComponent.tsx for consistency
- Renamed file to match the exported component name
- Updated import in index.ts barrel export
- Maintains consistency with other component naming patterns
