+++ archon-ui-main/src/features/projects/tasks/components/index.ts
@@ -13,4 +13,4 @@ export type { TaskCardProps } from "./TaskCard";
 export { TaskCard } from "./TaskCard";
 export { TaskCardActions } from "./TaskCardActions";
 export { TaskEditModal } from "./TaskEditModal";
-export { TaskPriorityComponent } from "./TaskPriority";
+export { TaskPriorityComponent } from "./TaskPriorityComponent";
